Move entry file to components folder
